feat: polish UI and refactor backend routes
