+++ frontend/src/App.js
@@ -10,7 +10,7 @@ import Home from './components/Home';
 import AccountLookup from './components/AccountLookup';
 import ValidatorList from './components/ValidatorList';
 import WalletPage from './components/WalletPage';
-import AdminPage from './components/AdminPage';
+import NotFound from './components/NotFound';
 import BlockDetails from './components/BlockDetails';
 import TransactionDetails from './components/TransactionDetails';
 import AccountPage from './components/AccountPage';
@@ -80,10 +80,11 @@ function App() {
         <Route path="/account" element={<div className="container"><AccountLookup /></div>} />
         <Route path="/validators" element={<div className="container"><ValidatorList /></div>} />
         <Route path="/wallet" element={<WalletPage wallet={wallet} login={login} logout={logout} />} />
-        <Route path="/admin" element={<AdminPage />} />
         <Route path="/block/:number" element={<BlockDetails />} />
         <Route path="/tx/:id" element={<TransactionDetails />} />
         <Route path="/account/:address" element={<AccountPage />} />
+        <Route path="/404" element={<NotFound />} />
+        <Route path="*" element={<NotFound />} />
       </Routes>
     </Router>
   );
