+++ frontend/src/components/AccountPage.js
@@ -11,9 +11,12 @@ function AccountPage() {
   const [account, setAccount] = useState(null);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(false);
+  const [page, setPage] = useState(1);
+  const perPage = 10;
 
   useEffect(() => {
     setLoading(true);
+    setPage(1);
     fetch(`${API_URL}/api/v1/accounts/${address}`)
       .then(res => {
         if (!res.ok) throw new Error('bad');
@@ -34,11 +37,21 @@ function AccountPage() {
   if (loading) return <Spinner />;
   if (error || !account) return <p>{t('Service unavailable')}</p>;
 
+  const total = Math.ceil(account.transactions.length / perPage);
+  const start = (page - 1) * perPage;
+  const end = start + perPage;
+  const txs = account.transactions.slice(start, end);
+
   return (
     <div className="container">
       <h2>{t('Account')} {address}</h2>
       <p>{t('Balance')}: {account.balance} TRI</p>
       <h4>{t('Transactions')}</h4>
+      <div className="pagination-controls">
+        <button onClick={() => setPage(p => Math.max(1, p - 1))} disabled={page === 1}>{t('Previous')}</button>
+        <span className="ml-sm">{t('Page')} {page} {t('of')} {total}</span>
+        <button onClick={() => setPage(p => p + 1)} disabled={page >= total} className="ml-sm">{t('Next')}</button>
+      </div>
       <table className="full-width">
         <thead>
           <tr>
@@ -50,7 +63,7 @@ function AccountPage() {
           </tr>
         </thead>
         <tbody>
-          {account.transactions.map(tx => (
+          {txs.map(tx => (
             <tr key={tx.txId}>
               <td><Link to={`/tx/${tx.txId}`}>{tx.txId}</Link></td>
               <td><Link to={`/account/${tx.from}`}>{tx.from}</Link></td>
